Fixed bugs that made the server crash on Linux

+++ bridge/resources/WiredPlayers/character/Inventory.cs
@@ -21,7 +21,7 @@ namespace WiredPlayers.character
             {
                 // Get the hash from the object
                 WeaponHash weaponHash = NAPI.Util.WeaponNameToModel(item.hash);
-                int hash = weaponHash == 0 ? int.Parse(item.hash) : (int)NAPI.Util.GetHashKey(Constants.WEAPON_ITEM_MODELS[weaponHash]);
+                uint hash = weaponHash == 0 ? uint.Parse(item.hash) : NAPI.Util.GetHashKey(Constants.WEAPON_ITEM_MODELS[weaponHash]);
 
                 // Create each of the items on the ground
                 item.objectHandle = NAPI.Object.CreateObject(hash, item.position, new Vector3(), 255, item.dimension);
